# api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ProtoAI MCP API",
    "version": "1.0.0",
    "description": "API para comunicação semântica entre AIs e APIs usando o Model Context Protocol"
  },
  "paths": {
    "/api/status": {
      "get": {
        "summary": "Verifica o status do servidor",
        "responses": {
          "200": {
            "description": "Status do servidor",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string" },
                    "version": { "type": "string" },
                    "timestamp": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/models": {
      "post": {
        "summary": "Registra um novo modelo",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "manifest": { "type": "object" },
                  "capabilities": {
                    "type": "array",
                    "items": { "type": "string" }
                  },
                  "metadata": { "type": "object" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Modelo registrado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "model_id": { "type": "string" },
                    "registration_time": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/discover": {
      "post": {
        "summary": "Descobre modelos compatíveis",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "intent": { "type": "string" },
                  "required_capabilities": {
                    "type": "array",
                    "items": { "type": "string" }
                  },
                  "constraints": { "type": "object" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Lista de modelos compatíveis",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "matches": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "model_id": { "type": "string" },
                          "compatibility_score": { "type": "number" },
                          "manifest": { "type": "object" },
                          "capability_mapping": { "type": "object" }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/connections": {
      "post": {
        "summary": "Estabelece uma conexão entre modelos",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "source_model_id": { "type": "string" },
                  "target_model_id": { "type": "string" },
                  "intent": { "type": "string" },
                  "connection_params": { "type": "object" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Conexão estabelecida com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "connection_id": { "type": "string" },
                    "status": { "type": "string" },
                    "connection_details": { "type": "object" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/models/{modelId}/connections": {
      "get": {
        "summary": "Lista conexões de um modelo",
        "parameters": [
          {
            "name": "modelId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de conexões",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "connections": {
                      "type": "array",
                      "items": { "type": "object" }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/connections/{connectionId}/interactions": {
      "post": {
        "summary": "Registra uma interação em uma conexão",
        "parameters": [
          {
            "name": "connectionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "interaction_type": { "type": "string" },
                  "request_data": { "type": "object" },
                  "response_data": { "type": "object" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Interação registrada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "interaction_id": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}